Move ceph eviction code from manila to manila-ganesha
This is so that manila *only* tests without ceph will not try to
clean-up non-existant ceph units.
